# docs/swagger.yaml
basePath: /
definitions:
  dto.AvailabilityAPIResponse:
    properties:
      data:
        $ref: '#/definitions/dto.AvailabilityData'
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.AvailabilityData:
    properties:
      available_date:
        example: "2026-08-31"
        type: string
      created_at:
        type: string
      end_time:
        example: "12:00"
        type: string
      id:
        example: 1
        type: integer
      published_at:
        type: string
      published_by:
        type: integer
      start_time:
        example: "08:00"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.AvailabilityStatus'
        example: PENDING
      trainer_id:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
  dto.AvailabilityListAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.AvailabilityData'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.ErrorResponse:
    properties:
      errors:
        type: object
      message:
        example: error message
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.ScheduleSlotData:
    properties:
      end_time:
        example: "10:00"
        type: string
      start_time:
        example: "08:00"
        type: string
    type: object
  dto.ScheduleTrainerData:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Pandu Pratama
        type: string
    type: object
...
  dto.StudentScheduleData:
    properties:
      availability_id:
        example: 1
        type: integer
      available_date:
        example: "2026-08-31"
        type: string
      slots:
        items:
          $ref: '#/definitions/dto.ScheduleSlotData'
        type: array
      trainer:
        $ref: '#/definitions/dto.ScheduleTrainerData'
    type: object
  dto.StudentScheduleListAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.StudentScheduleData'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.UpsertAvailabilityRequest:
    properties:
      available_date:
        example: "2026-08-31"
        type: string
      end_time:
        example: "12:00"
        type: string
      start_time:
        example: "08:00"
        type: string
    required:
    - available_date
    - end_time
    - start_time
    type: object
...
  models.AvailabilityStatus:
    enum:
    - PENDING
    - PUBLISHED
    - CANCELLED
    type: string
    x-enum-varnames:
    - AvailabilityPending
    - AvailabilityPublished
    - AvailabilityCancelled
...
info:
  contact: {}
  description: REST API Final Project Bootcamp Intensif Golang
  title: Driving Course Management System API
  version: "1.0"
paths:
...
  /api/v1/admin/trainer-availabilities:
    get:
      description: List all trainer availability ranges. Requires an ACTIVE ADMIN
        account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvailabilityListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all trainer availability ranges
      tags:
      - Admin Trainer Availability
  /api/v1/admin/trainer-availabilities/{id}:
    get:
      description: Get trainer availability range. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Availability ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvailabilityAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get trainer availability range
      tags:
      - Admin Trainer Availability
  /api/v1/admin/trainer-availabilities/{id}/cancel:
    patch:
      description: Cancel trainer availability. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Availability ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvailabilityAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel trainer availability
      tags:
      - Admin Trainer Availability
  /api/v1/admin/trainer-availabilities/{id}/publish:
    post:
      description: Publish pending trainer availability. Requires an ACTIVE ADMIN
        account.
      parameters:
      - description: Availability ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvailabilityAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Publish pending trainer availability
      tags:
      - Admin Trainer Availability
...
  /api/v1/student/schedules:
    get:
      description: List published trainer schedules and available two-hour slots.
        Requires an ACTIVE STUDENT account.
      parameters:
      - description: Filter by YYYY-MM-DD
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StudentScheduleListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List published trainer schedules and available two-hour slots
      tags:
      - Student Schedules
...
  /api/v1/trainer/availabilities:
    get:
      description: List own availability ranges. Requires an ACTIVE TRAINER account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvailabilityListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List own availability ranges
      tags:
      - Trainer Availability
    post:
      consumes:
      - application/json
      description: Create own pending availability. Requires an ACTIVE TRAINER account.
      parameters:
      - description: Availability range
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertAvailabilityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AvailabilityAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create own pending availability
      tags:
      - Trainer Availability
  /api/v1/trainer/availabilities/{id}:
    get:
      description: Get own availability range. Requires an ACTIVE TRAINER account.
      parameters:
      - description: Availability ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvailabilityAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get own availability range
      tags:
      - Trainer Availability
    put:
      consumes:
      - application/json
      description: Update own pending availability. Requires an ACTIVE TRAINER account.
      parameters:
      - description: Availability ID
        in: path
        name: id
        required: true
        type: integer
      - description: Availability range
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertAvailabilityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvailabilityAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update own pending availability
      tags:
      - Trainer Availability
  /api/v1/trainer/availabilities/{id}/cancel:
    patch:
      description: Cancel own availability. Requires an ACTIVE TRAINER account.
      parameters:
      - description: Availability ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvailabilityAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel own availability
      tags:
      - Trainer Availability
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'Enter the token with the Bearer prefix, for example: "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"